slit candy plugins into separate files
