+++ lua/myplugins/candy/colors.lua
@@ -1,0 +1,4 @@
+return {
+    require 'myplugins.candy.colors.nvim-colorizer',
+    require 'myplugins.candy.colors.colorutils',
+}
