+++ lua/myplugins/candy/colors/colorutils.lua
@@ -1,0 +1,6 @@
+return { "max397574/colortils.nvim", disable=true,
+    cmd = "Colortils",
+    config = function()
+        require("colortils").setup()
+    end,
+}
