+++ lua/myplugins/candy/colors/nvim-colorizer.lua
@@ -1,0 +1,7 @@
+return { 'NvChad/nvim-colorizer.lua', disable=false,
+    config = function()
+        require'colorizer'.setup{
+            mode = ({'foreground','background', 'virtualtext'})[3],
+        }
+    end
+}
