+++ lua/myplugins/candy/fold.lua
@@ -1,0 +1,4 @@
+return {
+    require 'myplugins.candy.fold.pretty-fold',
+    require 'myplugins.candy.fold.fold-preview',
+}
