+++ lua/myplugins/candy/fold/fold-preview.lua
@@ -1,0 +1,6 @@
+return { 'anuvyklack/fold-preview.nvim', disable=false,
+    requires = 'anuvyklack/keymap-amend.nvim',
+    config = function()
+        require('fold-preview').setup()
+    end
+}
