+++ lua/myplugins/candy/ghost.lua
@@ -1,0 +1,4 @@
+return {
+    require 'myplugins.candy.ghost.vim-ghost',
+    require 'myplugins.candy.ghost.nvim-ghost',
+}
