+++ lua/myplugins/candy/leap.lua
@@ -1,0 +1,5 @@
+return {
+    require 'myplugins.candy.leap.leap',
+    require 'myplugins.candy.leap.eyeline',
+    require 'myplugins.candy.leap.pounce',
+}
