+++ lua/myplugins/candy/leap/eyeline.lua
@@ -1,0 +1,7 @@
+return { 'jinh0/eyeliner.nvim', disable=true,
+    config = function()
+        require'eyeliner'.setup {
+            highlight_on_key = true
+        }
+    end
+}
