+++ lua/myplugins/candy/nui.lua
@@ -1,0 +1,1 @@
+return { 'MunifTanjim/nui.nvim', disable=false }
